Refactoring: - split actions.js file into two files - extract findElement function from Actions class
