+++ lib/browser/actions/actions.js
@@ -5,42 +5,10 @@ var q = require('q'),
     inherit = require('inherit'),
     wd = require('wd'),
     promiseUtil = require('q-promise-utils'),
-    StateError = require('../errors/state-error'),
-    find = require('../find-func').find;
+    StateError = require('../../errors/state-error'),
+    find = require('../../find-func').find;
 
-function isInvalidMouseButton(button) {
-    if (typeof button !== 'undefined' && [0, 1, 2].indexOf(button) === -1) {
-        throw new TypeError('Mouse button should be 0 (left), 1 (right) or 2 (middle)');
-    }
-}
-
-function isInvalidKeys(keys) {
-    if (typeof keys !== 'string') {
-        if (!Array.isArray(keys)) {
-            return true;
-        }
-        return keys.some(function(chunk) {
-            return typeof chunk !== 'string';
-        });
-    }
-    return false;
-}
-
-function isInvalidElement(element) {
-    return (!element || !element._selector) && typeof element !== 'string';
-}
-
-function replaceStack(error, stack) {
-    var message = error.stack.split('\n')[0];
-    error.stack = [message].concat(stack.split('\n').slice(1)).join('\n');
-}
-
-function serializeFunc(func) {
-    return '(' + func.toString()  + '(window));';
-}
-
-var Actions = inherit({
-
+module.exports = inherit({
     __constructor: function(browser) {
         this._browser = browser;
         this._driver = browser._browser;
@@ -145,7 +113,7 @@ var Actions = inherit({
         var _this = this;
 
         this._pushAction(this.click, function mouseDown() {
-            return _this._findElement(element)
+            return findElement(element, _this._browser)
                 .then(function(element) {
                     return _this._driver.moveTo(element);
                 })
@@ -167,7 +135,7 @@ var Actions = inherit({
 
         var _this = this;
         this._pushAction(this.doubleClick, function() {
-            return _this._findElement(element)
+            return findElement(element, _this._browser)
                 .then(function() {
                     return _this._driver.moveTo(element);
                 })
@@ -203,7 +171,7 @@ var Actions = inherit({
 
         var _this = this;
         this._pushAction(this.mouseDown, function mouseDown() {
-            return _this._findElement(element)
+            return findElement(element, _this._browser)
                 .then(function(element) {
                     return _this._driver.moveTo(element);
                 })
@@ -223,9 +191,10 @@ var Actions = inherit({
         this._pushAction(this.mouseUp, function mouseUp() {
             var action;
             if (element) {
-                action = _this._findElement(element).then(function(element) {
-                    return _this._driver.moveTo(element);
-                });
+                action = findElement(element, _this._browser)
+                    .then(function(element) {
+                        return _this._driver.moveTo(element);
+                    });
             } else {
                 action = q.resolve();
             }
@@ -253,12 +222,13 @@ var Actions = inherit({
         }
 
         this._pushAction(this.mouseMove, function mouseMove() {
-            return _this._findElement(element).then(function(element) {
-                if (offset) {
-                    return _this._driver.moveTo(element, offset.x, offset.y);
-                }
-                return _this._driver.moveTo(element);
-            });
+            return findElement(element, _this._browser)
+                .then(function(element) {
+                    if (offset) {
+                        return _this._driver.moveTo(element, offset.x, offset.y);
+                    }
+                    return _this._driver.moveTo(element);
+                });
         });
 
         return this;
@@ -278,9 +248,10 @@ var Actions = inherit({
             }
 
             action = function sendKeys() {
-                return _this._findElement(element).then(function(element) {
-                    return _this._driver.type(element, keys);
-                });
+                return findElement(element, _this._browser)
+                    .then(function(element) {
+                        return _this._driver.type(element, keys);
+                    });
             };
         }
 
@@ -308,7 +279,7 @@ var Actions = inherit({
                     if (!isFile) {
                         return q.reject(new StateError(path + ' should be existing file'));
                     }
-                    return _this._findElement(element);
+                    return findElement(element, _this._browser);
                 })
                 .then(function(element) {
                     return element.sendKeys(path);
@@ -333,7 +304,7 @@ var Actions = inherit({
         var _this = this;
 
         this._pushAction(this.tap, function tap() {
-            return _this._findElement(element)
+            return findElement(element, _this._browser)
                 .then(function(elem) {
                     return _this._driver.tapElement(elem);
                 });
@@ -349,7 +320,7 @@ var Actions = inherit({
         var _this = this;
         this._pushAction(this.flick, function flick() {
             if (element) {
-                return _this._findElement(element)
+                return findElement(element, _this._browser)
                   .then(function(elem) {
                       return _this._driver.flick(elem, offsets.x, offsets.y, speed);
                   });
@@ -380,61 +351,66 @@ var Actions = inherit({
         return this;
     },
 
-    _findElement: function(element) {
-        if (element._element) {
-            return q.resolve(element._element);
-        }
-
-        if (typeof element === 'string') {
-            element = find(element);
-        }
-
-        return this._browser.findElement(element._selector)
-            .then(function(foundElement) {
-                Object.defineProperty(element, '_element', {
-                    value: foundElement,
-                    writable: false,
-                    enumerable: false
-                });
-
-                return foundElement;
-            })
-            .fail(function(error) {
-                if (error.status === 7) {
-                    return q.reject(new StateError('Could not find element with css selector in actions chain: '
-                        + error.selector));
-                }
-                return q.reject(error);
-            });
-    },
-
     perform: function() {
         return promiseUtil.sequence(this._actions);
     }
 });
 
-module.exports = inherit(Actions, {
-    __constructor: function(browser) {
-        this.__base(browser);
-        this._postActions = new Actions(browser);
-    },
-
-    getPostActions: function() {
-        return this._postActions._actions.length > 0 && this._postActions;
-    },
-
-    setWindowSize: function(width, height) {
-        var _this = this;
-        this._pushAction(this.setWindowSize, function setWindowSize() {
-            return _this._driver.getWindowSize()
-                .then(function(size) {
-                    if (size.width !== width || size.height !== height) {
-                        _this._postActions.setWindowSize(size.width, size.height);
-                    }
-                    return _this._driver.setWindowSize(width, height);
-                });
-        });
-
-        return this;
+function isInvalidMouseButton(button) {
+    if (typeof button !== 'undefined' && [0, 1, 2].indexOf(button) === -1) {
+        throw new TypeError('Mouse button should be 0 (left), 1 (right) or 2 (middle)');
     }
-});
+}
+
+function isInvalidKeys(keys) {
+    if (typeof keys !== 'string') {
+        if (!Array.isArray(keys)) {
+            return true;
+        }
+        return keys.some(function(chunk) {
+            return typeof chunk !== 'string';
+        });
+    }
+    return false;
+}
+
+function isInvalidElement(element) {
+    return (!element || !element._selector) && typeof element !== 'string';
+}
+
+function replaceStack(error, stack) {
+    var message = error.stack.split('\n')[0];
+    error.stack = [message].concat(stack.split('\n').slice(1)).join('\n');
+}
+
+function serializeFunc(func) {
+    return '(' + func.toString()  + '(window));';
+}
+
+function findElement(element, browser) {
+    if (element._element) {
+        return q.resolve(element._element);
+    }
+
+    if (typeof element === 'string') {
+        element = find(element);
+    }
+
+    return browser.findElement(element._selector)
+        .then(function(foundElement) {
+            Object.defineProperty(element, '_element', {
+                value: foundElement,
+                writable: false,
+                enumerable: false
+            });
+
+            return foundElement;
+        })
+        .fail(function(error) {
+            if (error.status === 7) {
+                return q.reject(new StateError('Could not find element with css selector in actions chain: '
+                    + error.selector));
+            }
+            return q.reject(error);
+        });
+}
