+++ lib/capture-session.js
@@ -3,11 +3,11 @@
 var util = require('util'),
     q = require('q'),
     inherit = require('inherit'),
-    find = require('../lib/find-func').find,
+    find = require('./find-func').find,
     _ = require('lodash'),
     logger = require('./utils').logger,
     debug = require('debug'),
-    StateError = require('../lib/errors/state-error');
+    StateError = require('./errors/state-error');
 
 module.exports = inherit({
     __constructor: function(browser) {
